requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
gunicorn
